Fix courses.view_count entity type: Long -> Integer
Hibernate's ddl-auto=validate caught it on the user's machine (this
sandbox has no Docker, so the earlier smoke tests never exercised
schema validation against a real MySQL): the DDL is
"view_count INT UNSIGNED NOT NULL DEFAULT 0", but Course.viewCount was
mapped as Long, which Hibernate expects as BIGINT.

    Schema validation: wrong column type encountered in column
    [view_count] in table [courses]; found [int unsigned
    (Types#INTEGER)], but expecting [bigint (Types#BIGINT)]

Changed Course.viewCount and every DTO that carries it
(CourseListRow/CourseListItem/CourseDetailResponse/CourseStatsResponse)
from Long to Integer to match. Re-audited every other entity's column
against V1__init.sql/V2__admin_additions.sql column-by-column
(including the native MySQL ENUM columns — organizations.type,
courses.type, course_enrollments.status — which validated fine
against the converter-based String mapping despite the type-name
difference, and every other INT/BIGINT/VARCHAR/DECIMAL/DATETIME
column) and found no other width or precision mismatch.

Hibernate's schema validator fails fast on the first bad column per
table (GroupedSchemaValidatorImpl / AbstractSchemaValidator.validateTable),
so this may not be the only one — please rerun `./gradlew test`.

+++ src/main/java/com/hackathonteam7/mainprojectbackend/course/Course.java
@@ -55,8 +55,9 @@ public class Course extends BaseTimeEntity {
     @Column(name = "is_ordered", nullable = false)
     private Boolean isOrdered;
 
+    /** DB 는 INT UNSIGNED (view_count) — BIGINT 가 아니다. */
     @Column(name = "view_count", nullable = false)
-    private Long viewCount;
+    private Integer viewCount;
 
     @Builder
     private Course(Organization organization, User creator, Reward reward, String name, String description,
@@ -69,7 +70,7 @@ public class Course extends BaseTimeEntity {
         this.type = type;
         this.status = status == null ? CourseStatus.DRAFT : status;
         this.isOrdered = isOrdered == null || isOrdered;
-        this.viewCount = 0L;
+        this.viewCount = 0;
     }
 
     public void updateDetails(String name, String description) {
